Move proposal parameters to the separate module
